requirements.txt
obsws-python
opencv-python
mediapipe
python-dotenv
